Demonstration of deep learning based CASTRA
Demonstration of deep learning based CASTRA on the device. The user verification module using gait runs on the device. The activity of the user (STILL, ON FOOT) is detected with confidence score. If the Activity is ON-FOOT, the app measures the trust score that the phone is with the right person. Initially, when the trust score is red, the phone is with the attacker. As soon as the phone is picked up by normal user, our algorithm detects it and increases the trust score. A new score is updated every 5-6 seconds and we can detect an adversarial gait in less than 6 seconds. Please contact us for the Deep learning (LSTM, CNN) and on-device implementation code.